RLIB-27: Semi-working state of rewriting everything to MPFR
GNU MPFR (http://www.mpfr.org) is a scientific library handling
arbitrary precision floating point numbers.

Internally, numbers were switched from fixed point values to mpfr_t.
The PDF documentation elaborates on why the switch was necessary.

struct rlib_value is now not exposed outside the library, and
its internal format also changed, which needs external users
(e.g the language binding modules) to be modified.

New APIs were introduced:
* rlib_value_alloc() - allocate a struct rlib_value
* rlib_value_init() - internal to the library, initialize
  struct rlib_value on the stack or part of a larger struct
* etc.

Modified internal and external APIs:
* rlib_free() is a void function, doesn't return 0.
* Lot of functions now need an "rlib *r" argument where they
  previously didn't.
* Functions that actually return boolean values now explicitly
  return gboolean instead of int or gint. Some functions
  changes from int 0/-1 to gboolean TRUE/FALSE.
* the input and output drivers' interface was sanitized so
  calls that don't need to return a value are now void.
* gfloat was replaced by gdouble.
* gint was replaced by gint64 in many cases internally.
  This was to make MPFR input and output better, intmax_t
  (from stdint.h) is equivalent to gint64 and this is the largest
  integer type that is expected to be 64-bit everywhere.
  TODO: This may be changed again somewhat, as there were a lot
  of automatic "sed" runs.
* struct rlib_value is now not exposed outside the library and
  its internal format also changed, which needs external users
  (e.g the language binding modules) to be modified.
* Lot of macros are modified to use accessor functions or don't
  exist now, that were using internal members of struct rlib_value
  which is now an opaque struct in rlib.h.
* Interfaces were cleaned up to avoid mixing GLIB and plain types.

All these made broke the previously established API and ABI,
so it is necessary to increase the library version which is
now 2.0.0.

Monetary formatting was re-implemented using FreeBSD's strfmon()
and this internal implementation also uses mpfr_t so RLIB doesn't
rely on the external availability of strfmon() anymore.

The library now compiles and some test examples work.

+++ libsrc/charencoder.c
@@ -18,20 +18,18 @@
  * Free Software Foundation, Inc., 59 Temple Place - Suite 330,
  * Boston, MA 02111-1307, USA.
  * 
- * $Id$s
- *
  * This module implements a characterencoder that converts a character
  * either to or from UTF8 which is the internal language of RLIB.
  * it is implemented as a simple 'class'.
- *
  */
+
+#include <config.h>
 
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
 #include <string.h>
 #include <locale.h>
-#include <config.h>
 #include <errno.h>
 
 #include "rlib-internal.h"
@@ -53,7 +51,7 @@ void rlib_charencoder_free(GIConv converter UNUSED) {
 #endif
 }
 
-gint rlib_charencoder_convert(GIConv converter UNUSED, gchar **inbuf, gsize *inbytes_left UNUSED, gchar **outbuf, gsize *outbytes_left UNUSED) {
+gint64 rlib_charencoder_convert(GIConv converter UNUSED, gchar **inbuf, gsize *inbytes_left UNUSED, gchar **outbuf, gsize *outbytes_left UNUSED) {
 #ifdef DISABLE_UTF8
 	/* The strlen is passed in here so we bump it by 1 */
 	*outbuf = g_strdup(*inbuf);
